+++ src/extensions/surveys/hooks/useContrastingTextColor.ts
@@ -3,12 +3,16 @@ import { SurveyAppearance } from '../../../posthog-surveys-types'
 import * as Preact from 'preact'
 import { getTextColor } from '../surveys-utils'
 
-export function useContrastingTextColor(options: { appearance: SurveyAppearance; defaultTextColor?: string }): {
+export function useContrastingTextColor(options: {
+    appearance: SurveyAppearance
+    defaultTextColor?: string
+    forceUpdate?: boolean
+}): {
     ref: Preact.RefObject<HTMLElement>
     textColor: string
 } {
     const ref = useRef<HTMLElement>(null)
-    const [textColor, setTextColor] = useState(options.defaultTextColor ?? 'white')
+    const [textColor, setTextColor] = useState(options.defaultTextColor ?? 'black')
 
     // TODO: useContext to get the background colors instead of querying the DOM
     useEffect(() => {
@@ -16,7 +20,7 @@ export function useContrastingTextColor(options: { appearance: SurveyAppearance;
             const color = getTextColor(ref.current)
             setTextColor(color)
         }
-    }, [options.appearance])
+    }, [options.appearance, options.forceUpdate])
 
     return {
         ref,
